requirements.txt
altair==5.4.1
asgiref==3.8.1
attrs==24.2.0
azure-ai-documentintelligence==1.0.0b2
azure-ai-formrecognizer==3.3.3
azure-ai-vision-imageanalysis==1.0.0b3
azure-common==1.1.28
azure-core==1.30.2
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Django==4.1.13
djongo==1.3.6
dnspython==2.7.0
gitdb==4.0.11
GitPython==3.1.43
idna==3.8
isodate==0.6.1
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
msrest==0.7.1
narwhals==1.6.0
numpy==2.1.0
oauthlib==3.2.2
packaging==24.1
pandas==2.2.2
pillow==10.4.0
protobuf==5.28.0
pyarrow==17.0.0
pydeck==0.9.1
Pygments==2.18.0
pymongo==4.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.8.0
rpds-py==0.20.0
six==1.16.0
smmap==5.0.1
sqlparse==0.2.4
tenacity==8.5.0
toml==0.10.2
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
watchdog==4.0.2
